+++ pkg/alertmanager/operator.go
@@ -275,6 +275,16 @@ func (c *Operator) bootstrap(ctx context.Context, config operator.Config) error 
 			c.kclient,
 			resyncPeriod,
 			func(options *metav1.ListOptions) {
+				// TODO(simonpasquier): use a more restrictive label selector
+				// selecting only Alertmanager statefulsets (e.g.
+				// "app.kubernetes.io/name in (alertmanager)").
+				//
+				// We need to wait for a couple of releases after [1] merges to
+				// ensure that the expected labels have been propagated to the
+				// Alertmanager statefulsets otherwise the informer won't
+				// select any object.
+				//
+				// [1] https://github.com/prometheus-operator/prometheus-operator/pull/7786
 				options.LabelSelector = operator.ManagedByOperatorLabelSelector()
 			},
 		),